Removed broken RemoveShortJumps
